Move doneSince to helpers
